% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nyc_lion_node_streets_df}
\alias{nyc_lion_node_streets_df}
\title{NYC LION node-to-street-name lookup}
\format{
\subsection{\code{nyc_lion_node_streets_df}}{

A tibble with 245,070 rows and 2 columns:
\describe{
\item{node_id}{Node identifier, joins to \link{nyc_lion_nodes_sf}
\code{nodeid} and to \link{nyc_lion_sf} \code{node_id_from} / \code{node_id_to}.}
\item{street_name}{Street name associated with the node.}
}
}
}
\source{
NYC Department of City Planning, LION
(\url{https://www.nyc.gov/site/planning/data-maps/open-data.page}).
}
\usage{
nyc_lion_node_streets_df
}
\description{
Lists every street name associated with each LION node. Because most
nodes are the meeting point of two or more segments, a given
\code{node_id} typically has multiple associated street names (one row
per name).
}
\details{
Edition 25C, "BYTES of the BIG APPLE", published 2025-08-18 by NYC
Department of City Planning. Update frequency is quarterly.
}
\author{
Kieran Healy
}
\keyword{datasets}
