% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nyc_lion_nodes_sf}
\alias{nyc_lion_nodes_sf}
\title{NYC LION nodes}
\format{
\subsection{\code{nyc_lion_nodes_sf}}{

A simple feature collection with 139,421 rows and 3 columns (point
geometry):
\describe{
\item{nodeid}{Unique node identifier.}
\item{v_intersect}{\code{"VirtualIntersection"} if the node is flagged
as a virtual intersection in DOITT's CSCL database; blank
otherwise.}
\item{Shape}{Point geometry in EPSG:2263.}
}
}
}
\source{
NYC Department of City Planning, LION
(\url{https://www.nyc.gov/site/planning/data-maps/open-data.page}).
}
\usage{
nyc_lion_nodes_sf
}
\description{
Point locations of every node (topological junction) in the LION
network. A node occurs wherever two or more LION linear features
cross, regardless of whether a physical intersection occurs there.
Joins to \link{nyc_lion_sf} via \code{node_id_from} / \code{node_id_to}, and to
\link{nyc_lion_node_streets_df} via \code{nodeid}.
}
\details{
EPSG:2263, NAD83 / New York Long Island (ftUS).

Edition 25C, "BYTES of the BIG APPLE", published 2025-08-18 by NYC
Department of City Planning. Update frequency is quarterly.
}
\author{
Kieran Healy
}
\keyword{datasets}
